requirements.txt
mysql-connector-python==8.0.20
pip==20.1.1
pyTelegramBotAPI==3.6.6
python-telegram-bot==11.1.0
requests==2.22.0
api==0.0.7
apiai==1.2.3
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7